Adds Ritual of Entropy
Implements the Ritual of Entropy, which converts items in a chest to cobblestone based on their crafting recipe value.

The ritual extracts items from a chest placed above the master ritual stone and inserts cobblestone based on a calculated value. LP is consumed per conversion.

Currently uses a simplified heuristic for determining item value due to changes in the recipe system in 1.20.1. Recipe tree walking will be implemented in a future update.