Adds Ritual of Culling
Adds a new ritual that kills entities in a specified range.

This ritual is powerful, enabling the killing of non-boss
entities and, with sufficient demon will, even bosses.
It also has the capability to destroy primed TNT based on
configuration settings.

The ritual uses a significant amount of LP on activation and
a smaller amount per entity killed. The ritual also attempts
to create destructive demon will.