Improves Sigil of Equivalency functionality
Adds scrolling functionality to adjust the Sigil of Equivalency's radius, providing more granular control over its area of effect.

Implements plane-based block selection, replacing blocks only on the same plane as the clicked face, resulting in more predictable and controlled block replacements.

Reduces the default radius in the config to 1, as the radius is now easily adjustable in-game.

Adds visual feedback to the block replacement process via particles and sound.