reBalanced Diabolical Fungi and Temporal Dominance
adjusts the Temporal Dominance sigil's refresh cost to 15000 LP instead of 1k.

These changes make the Diabolical Fungi a more viable mana generation method
and rebalances the Temporal Dominance sigil to be more in line with its
power.

Fixes an issue where the Sigil of Repare would lose its active status.
Now it updates the ItemStack reference to avoid this.