Considers bound Key of Binding for Blood Apple LP
When a player consumes a Blood Apple, this change checks for a bound Key of Binding in their inventory or Curios slots.

If a bound key is found, the LiP generated by the apple are added to the soul network of the key's owner, rather than the player consuming the apple.

This allows players to contribute to another player's network by consuming the Blood Apple.