Added Sigil of the Demon Monk and Endless Greed Ritual
Sigil of the Demon Monk is a curio sigil that enhances unarmed combat, mining, and defense, scaling with demon will. It also features an execute mechanic, projectile catching, and kill bonuses.

Added the Ritual of Endless Greed, which collects mob drops and XP orbs into a container placed above the ritual stone. Should be very performant and server friendly as it deletes any drops it doesnt have room for instead of spilling in world.

Also includes various configuration options for both additions.