Updated to version 3.0.10, mostly bug fixes
Updated the mod version to 3.0.10 and incorporated several bug fixes.

This includes fixes for LP casting in the Blood-Infused Spellbook,
compat runes not appearing in creative tab and being unmineable, and
mixed mineable tags not being set correctly for runes.

Additionally, it adds enhancements to the Sanguine Diviner,
allowing users to right-click on Arcane Runes and Runes of
Unleashed Nature to view their status. It also displays a rune breakdown
on the Blood Altar.

Includes the Mixin annotation processor for compile-time validation and handles duplicate resources.